+++ semana18/cookenu/src/data/recipe/selectAllRecipes.ts
@@ -1,12 +1,11 @@
-import { connection } from "./connection/connection"
+import { connection } from "../connection/connection"
 
-export const selectRecipeById = async(id: string): Promise<any> => {
+export const selectAllRecipes = async(): Promise<any> => {
     try {
         const result = await connection('Recipes')
       .select("*")
-      .where({ id })
-
-    return result[0]
+      .orderBy('created_at', 'desc')
+    return result
     } catch (error) {
         throw new Error(error.message || error.sqlMessage);
     }
