+++ semana18/cookenu/src/data/recipe/selectRecipeById.ts
@@ -1,11 +1,12 @@
-import { connection } from "./connection/connection"
+import { connection } from "../connection/connection"
 
-export const selectAllRecipes = async(): Promise<any> => {
+export const selectRecipeById = async(id: string): Promise<any> => {
     try {
         const result = await connection('Recipes')
       .select("*")
-      .orderBy('created_at', 'desc')
-    return result
+      .where({ id })
+
+    return result[0]
     } catch (error) {
         throw new Error(error.message || error.sqlMessage);
     }
