+++ semana18/cookenu/src/data/user/selectAllUsers.ts
@@ -1,12 +1,11 @@
-import { connection } from "./connection/connection"
+import { connection } from "../connection/connection"
 
-export const selectUserById = async(id: string): Promise<any> => {
+export const selectAllUsers = async(): Promise<any> => {
     try {
         const result = await connection('Users')
       .select("*")
-      .where({ id })
 
-    return result[0]
+    return result
     } catch (error) {
         throw new Error(error.message || error.sqlMessage);
     }
